# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BAHI",
    "description": "The udhaar book both sides can see. All data is synthetic.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/shops/{shop_id}/munshi": {
      "post": {
        "tags": [
          "munshi"
        ],
        "summary": "Talk Text",
        "description": "He typed to the munshi.",
        "operationId": "talk_text_shops__shop_id__munshi_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MunshiIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MunshiOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/munshi/voice": {
      "post": {
        "tags": [
          "munshi"
        ],
        "summary": "Talk Voice",
        "description": "He spoke to the munshi. The book's names are Saaras's hints.",
        "operationId": "talk_voice_shops__shop_id__munshi_voice_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_talk_voice_shops__shop_id__munshi_voice_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MunshiOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/munshi/{conversation_id}/cards/{draft_id}/yes": {
      "post": {
        "tags": [
          "munshi"
        ],
        "summary": "Card Yes",
        "description": "He tapped हाँ, or the three-second countdown ran out. Saving twice is saving\nonce.",
        "operationId": "card_yes_shops__shop_id__munshi__conversation_id__cards__draft_id__yes_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          },
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MunshiOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/munshi/{conversation_id}/cards/{draft_id}/no": {
      "post": {
        "tags": [
          "munshi"
        ],
        "summary": "Card No",
        "description": "He tapped नहीं: the card goes, nothing is written.",
        "operationId": "card_no_shops__shop_id__munshi__conversation_id__cards__draft_id__no_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          },
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MunshiOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/munshi/{conversation_id}/say/{turn_id}.wav": {
      "get": {
        "tags": [
          "munshi"
        ],
        "summary": "Say Reply",
        "description": "The munshi's reply in Sarvam's voice (Bulbul, shreya). 503 when voice is\noffline; the screen then shows the reply without saying it.",
        "operationId": "say_reply_shops__shop_id__munshi__conversation_id__say__turn_id__wav_get",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          },
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Turn Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/people/{person_id}/udhaar": {
      "get": {
        "tags": [
          "customer"
        ],
        "summary": "My Udhaar",
        "description": "B2. What he owes across every shop, each entry, and invitations waiting.",
        "operationId": "my_udhaar_people__person_id__udhaar_get",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyUdhaarOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/pay": {
      "post": {
        "tags": [
          "customer"
        ],
        "summary": "Pay",
        "description": "B3. He paid this shop everything he owes it, by UPI. Paytm moves the\nmoney; the book records which entries it paid, and the shop is told.",
        "operationId": "pay_shops__shop_id__pay_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaidOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/invite/accept": {
      "post": {
        "tags": [
          "customer"
        ],
        "summary": "Accept",
        "description": "He said yes to the shop's invite, on his own phone.",
        "operationId": "accept_shops__shop_id__invite_accept_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/invite/decline": {
      "post": {
        "tags": [
          "customer"
        ],
        "summary": "Decline",
        "operationId": "decline_shops__shop_id__invite_decline_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_talk_voice_shops__shop_id__munshi_voice_post": {
        "properties": {
          "audio": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Audio"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "audio"
        ],
        "title": "Body_talk_voice_shops__shop_id__munshi_voice_post"
      },
...
      "CardOut": {
        "properties": {
          "draft_id": {
            "type": "string",
            "format": "uuid",
            "title": "Draft Id"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "amount_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Paise"
          },
          "kind": {
            "type": "string",
            "enum": [
              "udhaar",
              "payment",
              "customer"
            ],
            "title": "Kind"
          },
          "new": {
            "type": "boolean",
            "title": "New"
          },
          "status": {
            "type": "string",
            "enum": [
              "shown",
              "saved",
              "replaced",
              "cancelled"
            ],
            "title": "Status"
          },
          "reasons": {
            "items": {
              "type": "string",
              "enum": [
                "weak_match",
                "large",
                "unusual",
                "new_customer"
              ]
            },
            "type": "array",
            "title": "Reasons"
          },
          "spoken_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spoken Text"
          },
          "entry_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Id"
          },
          "on_bahi": {
            "type": "boolean",
            "title": "On Bahi"
          }
        },
        "type": "object",
        "required": [
          "draft_id",
          "customer_id",
          "display_name",
          "tag",
          "amount_paise",
          "kind",
          "new",
          "status",
          "reasons",
          "spoken_text",
          "entry_id",
          "on_bahi"
        ],
        "title": "CardOut",
        "description": "The entry the munshi proposed, as the card shows it. The amount is the one\nstored on the draft, never the munshi's sentence."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InviteOut": {
        "properties": {
          "shop": {
            "$ref": "#/components/schemas/ShopOut"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "invited_at": {
            "type": "string",
            "format": "date-time",
            "title": "Invited At"
          }
        },
        "type": "object",
        "required": [
          "shop",
          "display_name",
          "invited_at"
        ],
        "title": "InviteOut"
      },
...
      "MunshiIn": {
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "text": {
            "type": "string",
            "maxLength": 400,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "MunshiIn",
        "description": "His words, typed. No conversation_id: a new conversation."
      },
      "MunshiOut": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Conversation Id"
          },
          "heard": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heard"
          },
          "source": {
            "type": "string",
            "enum": [
              "typed",
              "sarvam",
              "sarvam_cached",
              "clip_script",
              "tap"
            ],
            "title": "Source"
          },
          "reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply"
          },
          "say_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Say Url"
          },
          "done": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Done"
          },
          "card": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CardOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished": {
            "type": "boolean",
            "title": "Finished"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "heard",
          "source",
          "reply",
          "say_url",
          "done",
          "card",
          "finished"
        ],
        "title": "MunshiOut"
      },
      "MyShopOut": {
        "properties": {
          "shop": {
            "$ref": "#/components/schemas/ShopOut"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "balance_paise": {
            "type": "integer",
            "title": "Balance Paise"
          },
          "day": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ThreadEntryOut"
            },
            "type": "array",
            "title": "Entries"
          },
          "unread": {
            "type": "integer",
            "title": "Unread"
          }
        },
        "type": "object",
        "required": [
          "shop",
          "customer_id",
          "display_name",
          "balance_paise",
          "day",
          "entries",
          "unread"
        ],
        "title": "MyShopOut"
      },
      "MyUdhaarOut": {
        "properties": {
          "today": {
            "type": "string",
            "format": "date",
            "title": "Today"
          },
          "person_id": {
            "type": "string",
            "title": "Person Id"
          },
          "total_paise": {
            "type": "integer",
            "title": "Total Paise"
          },
          "shops": {
            "items": {
              "$ref": "#/components/schemas/MyShopOut"
            },
            "type": "array",
            "title": "Shops"
          },
          "invites": {
            "items": {
              "$ref": "#/components/schemas/InviteOut"
            },
            "type": "array",
            "title": "Invites"
          }
        },
        "type": "object",
        "required": [
          "today",
          "person_id",
          "total_paise",
          "shops",
          "invites"
        ],
        "title": "MyUdhaarOut"
      },
...
      "PaidOut": {
        "properties": {
          "shop": {
            "$ref": "#/components/schemas/ShopOut"
          },
          "amount_paise": {
            "type": "integer",
            "title": "Amount Paise"
          },
          "paid_at": {
            "type": "string",
            "format": "date-time",
            "title": "Paid At"
          },
          "method": {
            "type": "string",
            "const": "upi",
            "title": "Method"
          },
          "settled_in": {
            "type": "integer",
            "title": "Settled In"
          },
          "entry_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entry Ids"
          },
          "elsewhere": {
            "items": {
              "$ref": "#/components/schemas/MyShopOut"
            },
            "type": "array",
            "title": "Elsewhere"
          }
        },
        "type": "object",
        "required": [
          "shop",
          "amount_paise",
          "paid_at",
          "method",
          "settled_in",
          "entry_ids",
          "elsewhere"
        ],
        "title": "PaidOut"
      },
      "PersonIn": {
        "properties": {
          "person_id": {
            "type": "string",
            "format": "uuid",
            "title": "Person Id"
          }
        },
        "type": "object",
        "required": [
          "person_id"
        ],
        "title": "PersonIn"
      },
...
      "ShopOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "locality": {
            "type": "string",
            "title": "Locality"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "locality"
        ],
        "title": "ShopOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}